Move components to `$lib/components`
