+++ client/src/lib/components/GoogleIcon.svelte
