+++ client/src/lib/components/IntersectionObserver.svelte
