+++ client/src/routes/MealDetailsView.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import Icon from '@iconify/svelte';
-	import Image from './Image.svelte';
-	import IntersectionObserver from './IntersectionObserver.svelte';
-  import Loading from './Loading.svelte';
+	import Image from '$lib/components/Image.svelte';
+	import IntersectionObserver from '$lib/components/IntersectionObserver.svelte';
+  import Loading from '$lib/components/Loading.svelte';
 
   export let meal: Meal;
   export let mealDetails: MealDetails;
